# Remove first result from measurement