## What this patch does to fix the issue.
This patch excludes first result from performance measurement because a first result includes some initialization time to read some data into memory.
That can be seen especially when we use pb file as model. 

Below is the compared result of `run.py` which we use a same pb file

* Before
INFO:__main__:Function _run took 0.3625333949457854 seconds to complete
INFO:__main__:Function _run took 0.007619671057909727 seconds to complete
INFO:__main__:Function _run took 0.007477258099243045 seconds to complete
INFO:__main__:Function _run took 0.0068569539580494165 seconds to complete
INFO:__main__:Function _run took 0.006867909105494618 seconds to complete
INFO:__main__:Function _run took 0.007176596904173493 seconds to complete
INFO:__main__:Function _run took 0.007123536895960569 seconds to complete
INFO:__main__:Function _run took 0.007711157901212573 seconds to complete
INFO:__main__:Function _run took 0.007052185945212841 seconds to complete
INFO:__main__:Function _run took 0.007207290967926383 seconds to complete

* After
INFO:__main__:Function _run took 0.00717798201367259 seconds to complete
INFO:__main__:Function _run took 0.009376114001497626 seconds to complete
INFO:__main__:Function _run took 0.01757289608940482 seconds to complete
INFO:__main__:Function _run took 0.005577495088800788 seconds to complete
INFO:__main__:Function _run took 0.0052312680054455996 seconds to complete
INFO:__main__:Function _run took 0.004525961121544242 seconds to complete
INFO:__main__:Function _run took 0.004832894075661898 seconds to complete
INFO:__main__:Function _run took 0.004257153021171689 seconds to complete
INFO:__main__:Function _run took 0.0041026161052286625 seconds to complete
INFO:__main__:Function _run took 0.004274927079677582 seconds to complete

## Link to any relevant issues or pull requests.
Fix #929